Controller e Services de Emprestimo.

+++ src/main/java/com/bibliotecaelo/service/LivroService.java
@@ -5,6 +5,7 @@ import java.util.UUID;
 import com.bibliotecaelo.converter.LivroDTOConverter;
 import com.bibliotecaelo.domain.Livro;
 import com.bibliotecaelo.dto.LivroDTO;
+import com.bibliotecaelo.repository.EmprestimoRepository;
 import com.bibliotecaelo.repository.LivroRepository;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.ValidationException;
@@ -17,10 +18,14 @@ public class LivroService {
 
     private final LivroRepository repository;
     private final LivroDTOConverter converter;
+    private final EmprestimoRepository emprestimoRepository;
 
-    public LivroService(LivroRepository repository, LivroDTOConverter converter) {
+
+    public LivroService(LivroRepository repository, LivroDTOConverter converter,
+            EmprestimoRepository emprestimoRepository) {
         this.repository = repository;
         this.converter = converter;
+        this.emprestimoRepository = emprestimoRepository;
     }
 
     public LivroDTO create(LivroDTO livroDTO) {
@@ -28,13 +33,11 @@ public class LivroService {
             throw new ValidationException("Já Existe uma Livro Cadastrado com este Título!");
         }
 
-        Livro novoLivro = converter.from(livroDTO);
-        return converter.to(repository.save(novoLivro));
+        return converter.to(repository.save(converter.from(livroDTO)));
     }
 
     public LivroDTO findById(UUID livroId) {
-        return converter.to(repository.findById(livroId).orElseThrow(
-                () -> new EntityNotFoundException("Livro Não Encontrado a partir do id informado!")));
+        return converter.to(buscaLivro(livroId));
     }
 
     public Page<LivroDTO> findAll(Pageable pageable) {
@@ -42,14 +45,21 @@ public class LivroService {
     }
 
     public LivroDTO update(LivroDTO livroDTO) {
-        final Livro livroToUpdate = repository.findById(livroDTO.getId()).orElseThrow(
-                () -> new EntityNotFoundException("Livro Não Encontrado a partir do id informado!"));
+        final Livro livroToUpdate = buscaLivro(livroDTO.getId());
 
         return converter.to(repository.saveAndFlush(converter.from(livroDTO, livroToUpdate)));
     }
 
     public void deleteById(UUID livroId) {
-        // TODO validar se o LIVRO tem registro de empréstimo, se tiver, não pode deixar deletar
+        if (emprestimoRepository.existsByLivroId(livroId)) {
+            throw new ValidationException("Livro Possui Empréstimo vinculado, portanto NÃO será excluído!");
+        }
+
         repository.deleteById(livroId);
+    }
+
+    private Livro buscaLivro(UUID livroId) {
+        return repository.findById(livroId).orElseThrow(
+                () -> new EntityNotFoundException("Livro Não Encontrado a partir do id informado!"));
     }
 }
